requirements.txt
colorlog==6.8.0
fastapi==0.109.0
jsons==1.6.3
langchain==0.0.352
openai==0.28.1
python-dotenv==1.0.0
starlette-context==0.3.6
termcolor==2.4.0
tiktoken==0.5.2
unidecode==1.3.7
uvicorn==0.27.0